Added currency formatting

+++ src/main/java/it/polimi/tiw/tiw_html_pure/Controller/Results.java
@@ -91,7 +91,7 @@ public class Results extends HttpServlet {
         ProductDAO productDAO = new ProductDAO(connection);
 
         //Get results of search
-        Map<Product, Double> risultati;
+        Map<Product, Integer> risultati;
 
         try{
             risultati = productDAO.getProductsFromQueryString(queryString);
